Using clib for deps

+++ deps/file2str/file2str.h
@@ -5,7 +5,7 @@
  * @param path File's path.
  * @param len Pointer to the length of the string returned
  * @return string contents of file; otherwise NULL on error */
-char *file2strl(const char *path, unsigned long *len);
+char *file2strl(const char *path, unsigned int *len);
 
 char *file2str(const char *path);
